[Pause] add starting pause menu
When not paused, there is a button in the top right of the game to pause it.
This triggers an overlay to appear that for now only contains a button to
resume the game and nothing else. "Pause" does not yet actually pause the
game, merely hides it from the user, e.g. the landlord can still be moved.

Addionally, Unity updated the list of tags to include "HappinessBar", which
appears to have been used but not added to the stored data.