WinLevel: add next button and clean up
Change to white text to be more visible, and add a button to advance to the
next level. For now only goes to the "WinGame" scene, eventually once we have
multiple levels it should advance to the next one in a way that allows use to
reuse the same scene between each level instead of needing to duplicate it.
We'll figure that out when we get there.

Additionally, add WinGame scene to the build settings.