[LoseLevel] add buttons and scene interactions
Add LoseLevelController.cs, which controls the logic. This has a public static
entry point, LaunchLoseLevelScene(), which is used when the current scene in
lost.

LevelHandler.cs is updated to use this. When called, the LoseLevel scene is
launched, with knowledge about what level was just lost. Then, that scene has
two buttons - to try the level again, or to quit. If trying again, the level
that was previously stored is loaded again.

Additionally, update MainScene to connect the tenant to the LevelHandler
script, not sure when or why that broke.