[Pause] use escape key to pause or resume
When the game is running, the first frame that the escape key is pressed, the
game will pause, and if pressed again, it will resume. This replaces the old
code in GameHandler.cs that was calling Application.Quit(), which didn't
appear to be working when testing. GameHandler.cs is deleted entirely since
that was the only code it included, and the game handler prefab is updated
accordingly.