[HappinessBar] refactor to use a single image
Instead of having multiple smiley faces based on the current happiness, use a
single face that changes expression from a smile to an angry frown.

As before, there are 4 levels used, corresponding to 4 of the 6 happiness
images available (1, 3, 4, and 6). Additionally, apply the HappinessBar.cs
script to the bar within the prefab, rather than just in the main scene,
in preparation for other scenes being added.