Add placeholder Level2 scene
Currently just includes a button that triggers winning the level, to verify
that the changes to WinLevelController.cs work. WinLevelController now has a
static entry point, like LoseLevelController, for handling launching the scene
with knowledge about the current state.

Additionally, LevelHandler.cs is updated to call the WinLevelController method
instead of directly loading the "WinLevel" scene.