[PauseHandler] clean up display, apply to prefab
Move the size and font changes for the pause button from the individual scenes
to the prefab, so that they don't need to be manually duplicated for each
level. Additionally, apply the same changes to the buttons within the pause
menu (resume/restart/quit) to have a consistent look and feel for buttons.