[Sounds] persist volume across scenes
In PauseController.cs, add a static value with the current slider value. At the
start of the scene, if the value is set it will be applied to the audio mixer
and used to update the visual value of the slider itself in the pause menu.
Whenever the volume changes the static value is updated.