[Pause] add pause menu to some non-level scenes
Added to scenes for the backstory, losing a level, wining a level, and winning
the overall game. For scenes that are not an actual game level, the "restart"
button is hidden. I couldn't get it to show up in the introduction or
instructions scenes for some reason, will be added to those later.