[Notifications] handle multiple potential messages
Instead of happiness bars or the bills script each setting the display text,
they instead tell the notification bar code whether the tenant is happy or not
(for happiness bars) of whether there is a bill pending or not (for the bills
script). The notification bar then combines these to report the apartments
where the tenants are unhappy and about to leave, as well as if there are
bills to be paid.

To be able to show the apartment numbers of the unhappy tenants, the happiness
bar is updated to also know the apartment number it is associated with, and
the uses in the different apartment prefabs are updated to set that number.

Additionally, the notification text now overflows instead of being truncated.