[WinGame] use new art background
Unfortunately, I was unable to get the pause menu to show up now that there is
an image for the background. I suspect this is caused by the same issue that
prevented the pause menu from working on the introduction and instructions
scenes, and will try to trouble shoot in a subsequent commit.