Actually fix tenant animation display
Not an issue with Z location, but with the X - instead of having both the
tenant position and then the art position set within the tenant, set the X
offset of the tenant to 0 and switch the art offset based on that, so that when
the tenant flips it flips in the correct location. While this does mean that
the collision for collecting rent is a bit apart from the actual tenant
display, for now that is preferable to the bad display of the tenants, and can
be addressed in a follow-up.